+++ lemniscate/src/main/java/com/vlad1m1r/lemniscate/roulette/BaseRouletteProgressView.java
@@ -2,6 +2,8 @@ package com.vlad1m1r.lemniscate.roulette;
 
 import android.content.Context;
 import android.content.res.TypedArray;
+import android.os.Parcel;
+import android.os.Parcelable;
 import android.util.AttributeSet;
 
 import com.vlad1m1r.lemniscate.base.BaseCurveProgressView;
@@ -28,9 +30,9 @@ public abstract class BaseRouletteProgressView extends BaseCurveProgressView {
     protected float mDistanceFromCenter = 1f;
 
     /**
-     * Curve will be drawn on interval  [0, 2*numberOfCycles*π] before repeating
+     * Curve will be drawn on interval  [0, 2*mNumberOfCycles*π] before repeating
      */
-    protected float numberOfCycles = 1;
+    protected float mNumberOfCycles = 1;
 
 
     public BaseRouletteProgressView(Context context) {
@@ -48,7 +50,7 @@ public abstract class BaseRouletteProgressView extends BaseCurveProgressView {
             setRadiusFixed(a.getFloat(R.styleable.RouletteCurveProgressView_radiusFixed, mRadiusFixed));
             setRadiusMoving(a.getFloat(R.styleable.RouletteCurveProgressView_radiusMoving, mRadiusMoving));
             setDistanceFromCenter(a.getFloat(R.styleable.RouletteCurveProgressView_distanceFromCenter, mDistanceFromCenter));
-            setNumberOfCycles(a.getFloat(R.styleable.RouletteCurveProgressView_numberOfCycles, numberOfCycles));
+            setNumberOfCycles(a.getFloat(R.styleable.RouletteCurveProgressView_numberOfCycles, mNumberOfCycles));
         } finally {
             a.recycle();
         }
@@ -83,15 +85,86 @@ public abstract class BaseRouletteProgressView extends BaseCurveProgressView {
     }
 
     public float getNumberOfCycles() {
-        return numberOfCycles;
+        return mNumberOfCycles;
     }
 
     public void setNumberOfCycles(float numberOfCycles) {
-        this.numberOfCycles = numberOfCycles;
+        this.mNumberOfCycles = numberOfCycles;
     }
 
     @Override
     public double getT(int i) {
-        return i*numberOfCycles*2*Math.PI/mPrecision;
+        return i* mNumberOfCycles *2*Math.PI/mPrecision;
+    }
+
+    @Override
+    public Parcelable onSaveInstanceState() {
+        Parcelable superState = super.onSaveInstanceState();
+        RouletteCurveSavedState ss = new RouletteCurveSavedState(superState);
+
+        ss.radiusFixed = this.mRadiusFixed;
+        ss.radiusMoving = this.mRadiusMoving;
+        ss.distanceFromCenter = this.mDistanceFromCenter;
+        ss.numberOfCycles = this.mNumberOfCycles;
+
+        return ss;
+    }
+
+    @Override
+    public void onRestoreInstanceState(Parcelable state) {
+        if(!(state instanceof RouletteCurveSavedState)) {
+            super.onRestoreInstanceState(state);
+            return;
+        }
+
+        RouletteCurveSavedState ss = (RouletteCurveSavedState)state;
+        super.onRestoreInstanceState(ss.getSuperState());
+        //end
+
+        setRadiusFixed(ss.radiusFixed);
+        setRadiusMoving(ss.radiusMoving);
+        setDistanceFromCenter(ss.distanceFromCenter);
+        setNumberOfCycles(ss.numberOfCycles);
+    }
+
+
+
+    static class RouletteCurveSavedState extends BaseCurveProgressView.BaseCurveSavedState {
+
+        float radiusFixed;
+        float radiusMoving;
+        float distanceFromCenter;
+        float numberOfCycles;
+
+        RouletteCurveSavedState(Parcelable superState) {
+            super(superState);
+        }
+
+        private RouletteCurveSavedState(Parcel in) {
+            super(in);
+            this.radiusFixed = in.readFloat();
+            this.radiusMoving = in.readFloat();
+            this.distanceFromCenter = in.readFloat();
+            this.numberOfCycles = in.readFloat();
+        }
+
+        @Override
+        public void writeToParcel(Parcel out, int flags) {
+            super.writeToParcel(out, flags);
+            out.writeFloat(this.radiusFixed);
+            out.writeFloat(this.radiusMoving);
+            out.writeFloat(this.distanceFromCenter);
+            out.writeFloat(this.numberOfCycles);
+        }
+
+        public static final Parcelable.Creator<RouletteCurveSavedState> CREATOR =
+                new Parcelable.Creator<RouletteCurveSavedState>() {
+                    public RouletteCurveSavedState createFromParcel(Parcel in) {
+                        return new RouletteCurveSavedState(in);
+                    }
+                    public RouletteCurveSavedState[] newArray(int size) {
+                        return new RouletteCurveSavedState[size];
+                    }
+                };
     }
 }