+++ lemniscate/src/main/java/com/vlad1m1r/lemniscate/roulette/EpitrochoidProgressView.java
@@ -9,7 +9,7 @@ import android.util.AttributeSet;
 
 public class EpitrochoidProgressView extends BaseRouletteProgressView {
 
-    // mRadiusFixed = 5, mRadiusMoving=3, mDistanceFromCenter=5, numberOfCycles = 3 to get pentagram
+    // mRadiusFixed = 5, mRadiusMoving=3, mDistanceFromCenter=5, mNumberOfCycles = 3 to get pentagram
 
     public EpitrochoidProgressView(Context context) {
         super(context);