+++ CHANGELOG.md
@@ -2,6 +2,13 @@ Change Log
 ==========
 
 
+Version 1.0.2 *(2017-01-24)*
+----------------------------
+
+ * Fix: Added OnSaveState for Roulette curves
+ * Fix: Precision is beeing saved OnSaveState for all curves
+
+
 Version 1.0.1 *(2017-01-23)*
 ----------------------------